requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
SQLAlchemy
PyMySQL
python-dotenv
elasticsearch